requirements.txt
streamlit
torch
transformers
python-docx
pypdf
plotly
